requirements-vercel.txt
fastapi==0.68.0
pydantic==1.8.2
google-generativeai==0.3.0
python-multipart==0.0.5
